Moved the web mvc request handler to the right module
